+++ packages/lb-annotation/src/constant/style.ts
@@ -77,6 +77,12 @@ export const BASE_ICON: { [a: number]: any } = {
 
 /** 无效色 - 红：rgba(255, 51, 51, 1) */
 export const INVALID_COLOR = 'rgba(255, 51, 51, 1)';
+export const INVALID_COLOR_RGBA = {
+  r: 255,
+  g: 51,
+  b: 51,
+  a: 1,
+};
 
 export const NULL_COLOR = 'rgba(204, 204, 204, 1)';