+++ packages/lb-components/src/components/pointCloudView/hooks/usePointCloudViews.ts
@@ -3,12 +3,18 @@
  * @author Glenfiddish <edwinlee0927@hotmail.com>
  * @createdate 2022-08-17
  */
-import { PointCloudAnnotation, PointCloud, MathUtils, getCuboidFromPointCloudBox } from '@labelbee/lb-annotation';
+import {
+  PointCloudAnnotation,
+  PointCloud,
+  MathUtils,
+  getCuboidFromPointCloudBox,
+} from '@labelbee/lb-annotation';
 import {
   IPointCloudBox,
   EPerspectiveView,
   PointCloudUtils,
   IPolygonPoint,
+  IPolygonData,
 } from '@labelbee/lb-utils';
 import { useContext } from 'react';
 import { PointCloudContext } from '../PointCloudContext';
@@ -169,6 +175,50 @@ const sideViewPolygon2PointCloud = (
   return newBoxParams;
 };
 
+const backViewPolygon2PointCloud = (
+  newPolygon: any,
+  originPolygon: any,
+  selectedPointCloudBox: IPointCloudBox,
+  pointCloudInstance: PointCloud,
+) => {
+  // Notice. The sort of polygon is important.
+  const [point1, point2, point3] = newPolygon.pointList;
+  const [op1, op2, op3] = originPolygon.pointList;
+
+  // 2D centerPoint => 3D x & z
+  const newCenterPoint = MathUtils.getLineCenterPoint([point1, point3]);
+  const oldCenterPoint = MathUtils.getLineCenterPoint([op1, op3]);
+
+  const offset = {
+    x: newCenterPoint.x - oldCenterPoint.x,
+    y: newCenterPoint.y - oldCenterPoint.y,
+  };
+
+  const offsetCenterPoint = {
+    x: offset.x,
+    y: 0, // Not be used.
+    z: newCenterPoint.y - oldCenterPoint.y,
+  };
+
+  // 2D height => 3D depth
+  const height = MathUtils.getLineLength(point1, point2);
+  const oldHeight = MathUtils.getLineLength(op1, op2);
+  const offsetHeight = height - oldHeight; // 3D depth
+
+  // 2D width => 3D width
+  const width = MathUtils.getLineLength(point2, point3);
+  const oldWidth = MathUtils.getLineLength(op2, op3);
+  const offsetWidth = width - oldWidth; // 3D width
+
+  let { newBoxParams } = pointCloudInstance.getNewBoxByBackUpdate(
+    offsetCenterPoint,
+    offsetWidth,
+    offsetHeight,
+    selectedPointCloudBox,
+  );
+
+  return newBoxParams;
+};
 /**
  * NewBox synchronize sideView
  * @param boxParams
@@ -350,7 +400,7 @@ export const usePointCloudViews = () => {
 
   const selectedPointCloudBox = selectedBox?.info;
 
-  if (!topViewInstance || !sideViewInstance) {
+  if (!topViewInstance || !sideViewInstance || !backViewInstance) {
     return {
       topViewAddBox: () => {},
       topViewSelectedChanged: () => {},
@@ -372,22 +422,22 @@ export const usePointCloudViews = () => {
     size,
     imgList,
     trackConfigurable,
+    zoom,
   }: {
     newPolygon: any;
     size: ISize;
     imgList: IFileItem[];
     trackConfigurable?: boolean;
+    zoom: number;
   }) => {
-    const newImgList = imgList as any[];
-
     const extraData = {
-      attribute: topViewInstance.pointCloud2dOperation.defaultAttribute ?? ''
+      attribute: topViewInstance.pointCloud2dOperation.defaultAttribute ?? '',
     };
 
     if (trackConfigurable === true) {
       Object.assign(extraData, {
         trackID: PointCloudUtils.getNextTrackID({
-          imgList: newImgList,
+          imgList: [], // Just calculate by the pointCloudBoxList in current page.
           extraBoxList: pointCloudBoxList,
         }),
       });
@@ -417,14 +467,17 @@ export const usePointCloudViews = () => {
 
     polygonOperation.setSelectedIDs([newPolygon.id]);
     setSelectedIDs(boxParams.id);
-    syncPointCloudViews(PointCloudView.Top, newPolygon, boxParams);
-    addPointCloudBox(boxParams);
+    const newPointCloudList = addPointCloudBox(boxParams);
+    syncPointCloudViews(PointCloudView.Top, newPolygon, boxParams, zoom, newPointCloudList);
     addHistory({ newBoxParams: boxParams });
   };
 
   /** Top-view selected changed and render to other view */
-  const topViewSelectedChanged = () => {
-    const boxParams = selectedBox?.info;
+  const topViewSelectedChanged = (
+    newSelectedBox?: IPointCloudBox,
+    newPointCloudList?: IPointCloudBox[],
+  ) => {
+    const boxParams = newSelectedBox ?? selectedBox?.info;
     const polygonOperation = topViewInstance?.pointCloud2dOperation;
 
     polygonOperation.setSelectedIDs(selectedIDs);
@@ -433,7 +486,7 @@ export const usePointCloudViews = () => {
     }
 
     const polygon = polygonOperation.selectedPolygon;
-    syncPointCloudViews(PointCloudView.Top, polygon, boxParams);
+    syncPointCloudViews(PointCloudView.Top, polygon, boxParams, undefined, newPointCloudList);
   };
 
   /**
@@ -444,7 +497,24 @@ export const usePointCloudViews = () => {
    */
   const viewUpdateBox = (newPolygon: any, originPolygon: any, fromView: string) => {
     if (selectedPointCloudBox) {
-      let newBoxParams = sideViewPolygon2PointCloud(
+      let transfer2PointCloud;
+      let newBoxParams;
+
+      // Switch the right function.
+      switch (fromView) {
+        case PointCloudView.Back:
+          transfer2PointCloud = backViewPolygon2PointCloud;
+          break;
+        case PointCloudView.Side:
+          transfer2PointCloud = sideViewPolygon2PointCloud;
+          break;
+
+        default:
+          transfer2PointCloud = sideViewPolygon2PointCloud;
+          break;
+      }
+
+      newBoxParams = transfer2PointCloud(
         newPolygon,
         originPolygon,
         selectedPointCloudBox,
@@ -453,8 +523,7 @@ export const usePointCloudViews = () => {
       // Update count
       if (mainViewInstance) {
         const { count } = mainViewInstance.getSensesPointZAxisInPolygon(
-          getCuboidFromPointCloudBox(newBoxParams)
-            .polygonPointList as IPolygonPoint[],
+          getCuboidFromPointCloudBox(newBoxParams).polygonPointList as IPolygonPoint[],
           [
             newBoxParams.center.z - newBoxParams.depth / 2,
             newBoxParams.center.z + newBoxParams.depth / 2,
@@ -467,8 +536,9 @@ export const usePointCloudViews = () => {
         };
       }
 
-      updateSelectedBox(newBoxParams);
-      syncPointCloudViews(fromView, newPolygon, newBoxParams);
+      const newPointCloudList = updateSelectedBox(newBoxParams);
+      syncPointCloudViews(fromView, newPolygon, newBoxParams, undefined, newPointCloudList);
+      return newPointCloudList;
     }
   };
 
@@ -485,10 +555,17 @@ export const usePointCloudViews = () => {
    * @param polygon
    * @param size
    */
-  const topViewUpdateBox = (polygon: any, size: ISize) => {
+  const topViewUpdateBox = (polygon: IPolygonData, size: ISize) => {
     // If the selected Object is Polygon.
     if (selectedPolygon) {
-      pushHistoryUnderUpdatePolygon(polygon);
+      /**
+       * Notice. The Polygon need to be converted to pointCloud coordinate system for storage.
+       */
+      const newPolygon = {
+        ...polygon,
+        pointList: polygon.pointList.map((v) => PointCloudUtils.transferCanvas2World(v, size)),
+      };
+      pushHistoryUnderUpdatePolygon(newPolygon);
       return;
     }
 
@@ -505,8 +582,14 @@ export const usePointCloudViews = () => {
         _.pickBy(newBoxParams, (v, k) => ['width', 'height', 'x', 'y']),
       );
 
-      updateSelectedBox(newBoxParams);
-      syncPointCloudViews(PointCloudView.Top, polygon, selectedPointCloudBox);
+      const newPointCloudBoxList = updateSelectedBox(newBoxParams);
+      syncPointCloudViews(
+        PointCloudView.Top,
+        polygon,
+        selectedPointCloudBox,
+        undefined,
+        newPointCloudBoxList,
+      );
     }
   };
 
@@ -516,8 +599,19 @@ export const usePointCloudViews = () => {
    * @param polygon
    * @param boxParams
    */
-  const syncPointCloudViews = (omitView: string, polygon: any, boxParams: IPointCloudBox) => {
+  const syncPointCloudViews = async (
+    omitView: string,
+    polygon: any,
+    boxParams: IPointCloudBox,
+    zoom?: number,
+    newPointCloudBoxList?: IPointCloudBox[],
+  ) => {
     const dataUrl = currentData?.url;
+
+    if (newPointCloudBoxList) {
+      // Wait for the mainPointCloudData.
+      await ptCtx.syncAllViewPointCloudColor(newPointCloudBoxList);
+    }
 
     const viewToBeUpdated = {
       [PointCloudView.Side]: () => {
@@ -538,8 +632,11 @@ export const usePointCloudViews = () => {
         viewToBeUpdated[key]();
       }
     });
+    if (zoom) {
+      mainViewInstance?.updateCameraZoom(zoom);
+    }
+
     mainViewGenBox(boxParams);
-    mainViewInstance?.highlightOriginPointCloud(boxParams);
   };
 
   const pointCloudBoxListUpdated = (newBoxes: IPointCloudBox[]) => {
@@ -547,12 +644,21 @@ export const usePointCloudViews = () => {
     mainViewInstance?.generateBoxes(newBoxes);
   };
 
-  const initPointCloud3d = () => {
+  const initPointCloud3d = (size: ISize) => {
     if (!mainViewInstance) {
       return;
     }
 
-    mainViewInstance.initPerspectiveCamera();
+    const orthographicParams = {
+      left: -size.width / 2,
+      right: size.width / 2,
+      top: size.height / 2,
+      bottom: -size.height / 2,
+      near: 100,
+      far: -100,
+    };
+
+    mainViewInstance.initOrthographicCamera(orthographicParams);
     mainViewInstance.initRenderer();
     mainViewInstance.render();
   };
@@ -585,6 +691,7 @@ export const usePointCloudViews = () => {
         mainViewInstance?.generateBox(v);
       });
 
+      ptCtx.syncAllViewPointCloudColor(boxParamsList);
       ptCtx.setPointCloudResult(boxParamsList);
       ptCtx.setPolygonList(polygonList);
     } else {