+++ packages/lb-components/src/store/annotation/types.ts
@@ -64,6 +64,7 @@ export interface AnnotationState {
   beforeRotate?: () => boolean;
 
   pointCloudLoading: boolean;
+  checkMode: boolean; // Judge current Mode is checkMode or not.
 }
 
 interface UpdateToolInstance {
@@ -168,6 +169,9 @@ interface CopyBackWordResult extends CommonActions {
   type: typeof ANNOTATION_ACTIONS.COPY_BACKWARD_RESULT;
 }
 
+interface InitAnnotationState extends CommonActions {
+  type: typeof ANNOTATION_ACTIONS.INIT_ALL_STATE;
+}
 interface BatchUpdateTrackID {
   type: typeof ANNOTATION_ACTIONS.BATCH_UPDATE_TRACK_ID,
   payload: {
@@ -204,4 +208,5 @@ export type AnnotationActionTypes =
   | CopyBackWordResult
   | UpdateOnSave
   | BatchUpdateTrackID
-  | BatchUpdateResultByTrackID;
+  | BatchUpdateResultByTrackID
+  | InitAnnotationState;